cuda-compatibility-workarounds: add recipe
to install a header file that makes nvcc compilations compatible with
the glibc 2.38 for aarch64 targets. The path to this header is added
as an -isystem option passed via -Xcompiler on nvcc invocations, so
it overrides the glibc default, bypassing references to gcc
built-in SIMD types that the cicc compiler doesn't define.

Signed-off-by: Matt Madison <matt@madison.systems>