tegra-common: set OLDEST_KERNEL so that glibc will restrict features
As of the scarthgap release, OLDEST_KERNEL is set in meta.conf to
version 5.15. This is newer than the version available to us for the
tegra machines in the Jetpack 4 releases. As such, some components
that use this variable (such as glibc) will be build assuming support
for kernel features we don't have.

Override the OLDEST_KERNEL to 4.9, to match the linux-tegra kernel.